requirements.txt
fastapi
uvicorn[standard]
gunicorn
# psycopg2-binary
python-dotenv
asyncpg